Rename PersistentStorageUtils -> PersistentStorageMacros
