+++ src/lib/support/BUILD.gn
@@ -77,7 +77,7 @@ static_library("support") {
     "LifetimePersistedCounter.h",
     "PersistedCounter.cpp",
     "PersistedCounter.h",
-    "PersistentStorageUtils.h",
+    "PersistentStorageMacros.h",
     "Pool.cpp",
     "Pool.h",
     "PrivateHeap.cpp",
