+++ src/lib/support/PersistentStorageMacros.h
